refac: Move terminal input handling into input.rs
